+++ halotools/mock_observables/two_point_clustering/__init__.py
@@ -2,7 +2,7 @@
 on galaxy/halo clustering, e.g., three-dimensional clustering
 `~halotools.mock_observables.tpcf`, projected clustering `~halotools.mock_observables.wp`,
 RSD multipoles `~halotools.mock_observables.tpcf_multipole`,
-galaxy-galaxy lensing `~halotools.mock_observables.delta_sigma`, and more.
+galaxy-galaxy lensing `~halotools.mock_observables.mean_delta_sigma`, and more.
 """
 from __future__ import absolute_import